requirements.txt
pandas
numpy
requests
sqlalchemy
psycopg2-binary
feast
python-dotenv